add Tagline
Illarion in a nutshell. No change in lay-out required. Above the header
so it doesn't interfere with the logo on a small resolution. Will help
potential players see what Illarion is about without reading the entire
blob of homepage text (people usually don't stay on a webpage for longer
than 10 seconds). The German version uses "mit Roleplay-Enforcement"
because it's less stiff and easier to search for than "mit dem
Schwerpunkt Rollenspiel"