FIX ISSUE0010857 Start Date of quests is removed
us/de_quests_edit.php took the maximum days of a month using the variable
days_in_month which used the function mktime(1,1,1,$month,0,$year) where
as 0 referenced the 0th day of the month which is (following the
documentation) the last day of the month before. Bug was solved by setting
the day-value to 1.

For example March was handeld as February (it had only 28 days)
and if a day with a higher number than 28 was saved it
was reverted to the default value of 1.