refs #11584 Nickname wiping for GM Chars
GM Chars are now able to wipe the nick names assigned to them as well
as the introductions they did. This way all characters forget this
character.

The function is enabled to all GM characters and is accessible in the
account system. The function is in the account system because is should
be usable for questers that are unable to use the GM Tool.